+++ weblate/accounts/forms.py
@@ -30,7 +30,7 @@ from weblate.accounts.utils import (
     get_all_user_mails,
     invalidate_reset_codes,
 )
-from weblate.auth.models import Group, User, get_auth_keys
+from weblate.auth.models import Group, User
 from weblate.lang.models import Language
 from weblate.logger import LOGGER
 from weblate.trans.defines import FULLNAME_LENGTH
@@ -173,8 +173,11 @@ class CommitForm(ProfileBaseForm):
     commit_email = forms.ChoiceField(
         label=gettext_lazy("Commit e-mail"),
         choices=[("", gettext_lazy("Use account e-mail address"))],
-        help_text=gettext_lazy("Choose commit e-mail from verified addresses."),
+        help_text=gettext_lazy(
+            "Used in version control commits. The address will stay in the repository forever once changes are commited by Weblate."
+        ),
         required=False,
+        widget=forms.RadioSelect,
     )
 
     class Meta:
@@ -197,10 +200,6 @@ class CommitForm(ProfileBaseForm):
         self.helper = FormHelper(self)
         self.helper.disable_csrf = True
         self.helper.form_tag = False
-        layout = ["commit_email"]
-        if "email" in get_auth_keys():
-            layout.append(Div(template="accounts/add-mail.html"))
-        self.helper.layout = Layout(*layout)
 
 
 class ProfileForm(ProfileBaseForm):
@@ -342,10 +341,13 @@ class UserForm(forms.ModelForm):
 
     username = UniqueUsernameField()
     email = forms.ChoiceField(
-        label=gettext_lazy("E-mail"),
-        help_text=gettext_lazy("Choose primary e-mail from verified addresses."),
+        label=gettext_lazy("Account e-mail"),
+        help_text=gettext_lazy(
+            "Used for e-mail notifications and as a commit e-mail if it is not not configured below."
+        ),
         choices=(("", ""),),
         required=True,
+        widget=forms.RadioSelect,
     )
     full_name = FullNameField()
 
@@ -364,10 +366,6 @@ class UserForm(forms.ModelForm):
         self.helper = FormHelper(self)
         self.helper.disable_csrf = True
         self.helper.form_tag = False
-        layout = ["username", "email", "full_name"]
-        if "email" in get_auth_keys():
-            layout.insert(2, Div(template="accounts/add-mail.html"))
-        self.helper.layout = Layout(*layout)
 
     @classmethod
     def from_request(cls, request):